+++ demo/Page.svelte
@@ -5,9 +5,9 @@ import { link, navigate, route } from '@koffeine/svelte-router';
 /** @type {{ title: string }} */
 const { title } = $props();
 
-// Query params
-const numbers = $derived([ 1, 2, 3 ].sort((a, b) => (a - b) * (route.query.order === 'desc' ? -1 : 1)));
-const query = $derived({ order: route.query.order === 'desc' ? 'asc' : 'desc' });
+// Search params
+const numbers = $derived([ 1, 2, 3 ].sort((a, b) => (a - b) * (route.searchParams.order === 'desc' ? -1 : 1)));
+const searchParams = $derived({ order: route.searchParams.order === 'desc' ? 'asc' : 'desc' });
 </script>
 
 <h2>Page: {title}</h2>
@@ -21,7 +21,7 @@ Numbers:
 </ul>
 
 <!-- Link navigation -->
-<a href={route.path} {@attach link({ query })}>Reverse numbers (add a new history entry)</a>
+<a href={link(route.pathname, { searchParams })}>Reverse numbers (add a new history entry)</a>
 
 <!-- API navigation -->
-<button type="button" onclick={() => navigate(route.path, { query, replace: true })}>Reverse numbers (replace the current history entry)</button>
+<button type="button" onclick={() => navigate(route.pathname, { searchParams, replace: true })}>Reverse numbers (replace the current history entry)</button>