+++ test/index.test.js
@@ -1,96 +1,139 @@
-import { describe, test, expect } from 'vitest';
-import { check, mock } from './utils.js';
+import { describe, expect, test } from 'vitest';
 import { init, navigate } from '../src/index.js';
+import { check, mock } from './utils.js';
 
-const baseUrl = '/';
+/** @param {import('../src/index.js').PublicRouteConfig[]} routeConfigs */
+const setup = async (routeConfigs, baseUrl = '/svelte-router') => {
+	history.replaceState(null, '', baseUrl); // Playwright fix
 
-history.replaceState(null, '', baseUrl); // Playwright fix
-
-await init(
-	[
-		{ path: '/', component: mock('Index') },
-		{ path: '/redirect', redirect: '/redirected' },
-		{ path: '/redirected', component: mock('Redirected') },
-		{ path: '/params/:param', component: mock('Params') },
-		{ path: '/optional-params/:param?', component: mock('OptionalParams') },
-		{ path: '/query-params', component: mock('QueryParams') },
-		{ path: '/wildcard/*', component: mock('Wildcard') }
-	],
-	{ baseUrl }
-);
+	await init(routeConfigs, { baseUrl });
+};
 
 
-test('should handle initial navigation', () => {
-	check({ path: '/', component: 'Index' });
+test('should handle initial navigation', async () => {
+	await setup([ { pathname: '/', component: mock('Index') } ]);
+
+	check({ component: 'Index', pathname: '/' });
 });
 
-test('should handle navigation with redirect', async () => {
-	await navigate('/redirect');
+test('should handle initial navigation when baseUrl ends with \'/\'', async () => {
+	await setup([ { pathname: '/', component: mock('Index') } ], '/svelte-router/');
 
-	check({ path: '/redirected', component: 'Redirected' });
+	check({ component: 'Index', pathname: '/' });
 });
 
-describe('params', () => {
-	test('should handle navigation with params in path', async () => {
-		await navigate('/params/value');
+test('should handle back navigation', async () => {
+	await setup([
+		{ pathname: '/', component: mock('Index') },
+		{ pathname: '/page', component: mock('Page') }
+	]);
 
-		check({ path: '/params/value', component: 'Params', params: { param: 'value' } });
+	check({ component: 'Index', pathname: '/' });
+
+	await navigate('/page');
+
+	check({ component: 'Page', pathname: '/page' });
+
+	await new Promise((resolve) => {
+		window.addEventListener('popstate', resolve, { once: true });
+		history.back();
 	});
 
-	test('should handle navigation with params in options', async () => {
-		await navigate('/params/:param', { params: { param: 'value' } });
+	check({ component: 'Index', pathname: '/' });
+});
 
-		check({ path: '/params/value', component: 'Params', params: { param: 'value' } });
-	});
+test('should ignore navigation to the same url', async () => {
+	await setup([ { pathname: '/', component: mock('Index') } ]);
+
+	const length = history.length;
+
+	await navigate('/');
+
+	expect(history.length).toBe(length);
+});
+
+test('should handle navigation with params', async () => {
+	await setup([
+		{ pathname: '/', component: mock('Index') },
+		{ pathname: '/params/:param', component: mock('Params') }
+	]);
+
+	await navigate('/params/value');
+
+	check({ component: 'Params', pathname: '/params/value', params: { param: 'value' } });
 });
 
 describe('optional params', () => {
-	test('should handle navigation with optional params in path', async () => {
+	test('should handle navigation with optional params', async () => {
+		await setup([
+			{ pathname: '/', component: mock('Index') },
+			{ pathname: '/optional-params/:param?', component: mock('OptionalParams') }
+		]);
+
 		await navigate('/optional-params/value');
 
-		check({ path: '/optional-params/value', component: 'OptionalParams', params: { param: 'value' } });
-	});
-
-	test('should handle navigation with optional params in options', async () => {
-		await navigate('/optional-params/:param', { params: { param: 'value' } });
-
-		check({ path: '/optional-params/value', component: 'OptionalParams', params: { param: 'value' } });
+		check({ component: 'OptionalParams', pathname: '/optional-params/value', params: { param: 'value' } });
 	});
 
 	test('should handle navigation with no optional params', async () => {
+		await setup([
+			{ pathname: '/', component: mock('Index') },
+			{ pathname: '/optional-params/:param?', component: mock('OptionalParams') }
+		]);
+
 		await navigate('/optional-params');
 
 		// eslint-disable-next-line no-undefined
-		check({ path: '/optional-params', component: 'OptionalParams', params: { param: undefined } });
-	});
-});
-
-describe('query params', () => {
-	test('should handle navigation with query params in path', async () => {
-		await navigate('/query-params?param=value');
-
-		check({ path: '/query-params', component: 'QueryParams', query: { param: 'value' } });
-	});
-
-	test('should handle navigation with query params in options', async () => {
-		await navigate('/query-params', { query: { param: 'value' } });
-
-		check({ path: '/query-params', component: 'QueryParams', query: { param: 'value' } });
-	});
-
-	test('should handle navigation with query params in path and options', async () => {
-		await navigate('/query-params?param1=pathValue&param2=pathValue', { query: { param2: 'optionsValue', param3: 'optionsValue' } });
-
-		check({ path: '/query-params', component: 'QueryParams', query: { param1: 'pathValue', param2: 'optionsValue', param3: 'optionsValue' } });
+		check({ component: 'OptionalParams', pathname: '/optional-params', params: { param: undefined } });
 	});
 });
 
 test('should handle wildcard navigation', async () => {
+	await setup([
+		{ pathname: '/', component: mock('Index') },
+		{ pathname: '/wildcard/*', component: mock('Wildcard') }
+	]);
+
 	await navigate('/wildcard/any/thing');
 
-	check({ path: '/wildcard/any/thing', component: 'Wildcard', params: { '*': 'any/thing' } });
+	check({ component: 'Wildcard', pathname: '/wildcard/any/thing', params: { '*': 'any/thing' } });
 });
 
-test('should throw error on unknown path', async () => {
-	await expect(navigate('/not-found')).rejects.toThrowError();
+test('should handle navigation with search params', async () => {
+	await setup([
+		{ pathname: '/', component: mock('Index') },
+		{ pathname: '/search-params', component: mock('SearchParams') }
+	]);
+
+	await navigate('/search-params', { searchParams: { param: 'value' } });
+
+	check({ component: 'SearchParams', pathname: '/search-params', searchParams: { param: 'value' } });
+});
+
+test('should handle navigation with redirect', async () => {
+	await setup([
+		{ pathname: '/', component: mock('Index') },
+		{ pathname: '/redirect', redirect: '/redirected' },
+		{ pathname: '/redirected', component: mock('Redirected') }
+	]);
+
+	await navigate('/redirect');
+
+	check({ component: 'Redirected', pathname: '/redirected' });
+});
+
+test('should throw error when no route found for pathname', async () => {
+	await setup([ { pathname: '/', component: mock('Index') } ]);
+
+	await expect(navigate('/not-found'))
+		.rejects
+		.toThrow('No route found for pathname \'/not-found\'');
+});
+
+test('should throw error when pathname is outside of baseUrl', async () => {
+	history.replaceState(null, '', '/'); // Playwright fix
+
+	await expect(init([ { pathname: '/', component: mock('Index') } ], { baseUrl: '/svelte-router' }))
+		.rejects
+		.toThrow('Pathname \'/\' is outside of baseUrl \'/svelte-router\'');
 });